\name{lincombfun}
\alias{lincombfun}
\title{linear combination function
}  
\description{This is a prestep function for \emph{momentfun}
}
\usage{
lincombfun(S,mycoef)
} 
\arguments{
  \item{S}{S is a list of sums output from the sum function}
\item{mycoef}{global coefficients for the linear combination function}
}  
\references{YH Zhou, G Mayhew, Z Sun, X Xu, F Zou, FA Wright, 2013 \emph{Space-time clustering and the permutation moments of quadratic forms}, Stat 2(1), 292-302}
\seealso{\code{\link{coef}}, \code{\link{sumfun}}, \code{\link{momentfun}}.}
\keyword{lincombfun}