\name{mycoef}
\alias{mycoef}
\docType{data}
\title{Global variables for several main functions.
}
\description{We need the global variables to generate the linear combination function. 
}
\usage{data(mycoef)}
\details{It gives all the global variables
}
\references{YH Zhou, G Mayhew, Z Sun, X Xu, F Zou, FA Wright, 2013 \emph{Space-time clustering and the permutation moments of quadratic forms}, Stat 2(1), 292-302}

