% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad-package.R
\docType{package}
\name{quad-package}
\alias{quad}
\alias{quad-package}
\title{quad: Quadrilateral Stuff for Rasters}
\description{
Quads as first class types with helpers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/quad}
  \item Report bugs at \url{https://github.com/hypertidy/quad/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
