% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwad-package.r, R/qwad.r
\docType{package}
\name{qwad}
\alias{qwad}
\alias{qwad-package}
\alias{qwad.character}
\title{qwad.}
\usage{
qwad(x, ...)

\method{qwad}{character}(x, varname)
}
\arguments{
\item{x}{file name for character method}

\item{varname}{variable name}
}
\value{
Raster
}
\description{
qwad.

Title
}

