% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwadmap.r
\name{qwadmap}
\alias{qwadmap}
\alias{qwadmap.SpatialLinesDataFrame}
\alias{qwadmap.SpatialPointsDataFrame}
\alias{qwadmap.SpatialPolygonsDataFrame}
\title{Remap an object to the space defined by coordinate arrays.}
\usage{
qwadmap(x, ...)

\method{qwadmap}{SpatialPolygonsDataFrame}(x, coords, crop = FALSE,
  lonlat = TRUE, ...)

\method{qwadmap}{SpatialLinesDataFrame}(x, coords, crop = FALSE,
  lonlat = TRUE, ...)

\method{qwadmap}{SpatialPointsDataFrame}(x, coords, crop = FALSE,
  lonlat = TRUE, ...)
}
\arguments{
\item{x}{object to transform to the grid space, e.g. a \code{\link[sp]{Spatial}} object}

\item{...}{unused}

\item{coords}{qwadcoords RasterStack}

\item{crop}{logical, if \code{TRUE} crop x to the extent of the boundary of the values in coords}

\item{lonlat}{logical, if \code{TRUE} check for need to back-transform to longitude/latitude and do it}
}
\value{
input object with coordinates transformed to space of the coords
}
\description{
Find the nearest-neighbour coordinates of `x` in the coordinate arrays of `coords`.
}
\details{
The input `coords` is a assumed to be a 2-layer RasterStack or RasterBrick and
using `nabor::knn` the nearest matching position of the coordinates of `x` is found in the grid space of `coords`. The
motivating use-case is the curvilinear longitude and latitude arrays of ROMS model output, but the original ROMS-specific
code is being generalized (other models include CMIP5/6, Bluelink, etc.).

Cropping is complicated more details . . .
}

