\name{sumfun}
\alias{sumfun}
\title{ Get Sum
}
\description{This is a prestep function for \emph{momentfun}
}
\usage{
sumfun(W)
}  
\arguments{
  \item{W}{mid step for generating linear combination function  
}
}
\references{YH Zhou, G Mayhew, Z Sun, X Xu, F Zou, FA Wright, 2013 \emph{Space-time clustering and the permutation moments of quadratic forms}, Stat 2(1), 292-302}
