\name{quad-package}
\alias{quad-package}
\alias{quad}
\docType{package} 
\title{This package gives the exact the first four moments of any quadratic form and its corresponding pvlaues.
}
\description{The Mantel and Knox space-time clustering statistics are popular tools to establish transmissibility of a disease and detect outbreaks. The most commonly used null distributional approximations may provide poor fits, and researchers often resort to direct sampling from the permutation distribution. However, the exact first four moments for these statistics are available, and Pearson distributional approximations are often effective. Thus, our first goals are to clarify the literature and make these tools more widely available. In addition, by rewriting terms in the statistics, we obtain the exact first four permutation moments for the most commonly used quadratic form statistics, which need not be positive definite.  The extension of this work to quadratic forms greatly expands the utility of density approximations for these problems, including for high-dimensional applications, where the statistics must be extreme in order to exceed stringent testing thresholds. }

 
\details{   
\tabular{ll}{
Package: \tab quad\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-05\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Yi-Hui Zhou

Maintainer: Yi-Hui Zhou <yihui_zhou@ncsu.edu>

}
\references{YH Zhou, G Mayhew, Z Sun, X Xu, F Zou, FA Wright, 2013 \emph{Space-time clustering and the permutation moments of quadratic forms}, Stat 2(1), 292-302}
\keyword{quad}
