% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad-package.R
\docType{package}
\name{quad-package}
\alias{quad}
\alias{quad-package}
\title{quad: Intermediate Forms of Raster Grids}
\description{
Raster grids and quads as first class types with helpers. Provides a low level API for generating mesh index and vertices from the simplest abstraction of raster grid, input dimension and (optionally) extent to generate components of meshes for downstream visualization and efficient coordinate transformation. API functions may be 'Linked To' in the R headers library way.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/quad}
  \item Report bugs at \url{https://github.com/hypertidy/quad/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
